signedness warning

+++ franken/malloc/malloc.c
@@ -116,7 +116,7 @@ union	overhead {
 #define MINALIGN 16
 static	union overhead *nextf[NBUCKETS];
 
-static	long pagesz;			/* page size */
+static	size_t pagesz;			/* page size */
 static	int pagebucket;			/* page size bucket */
 
 #ifdef MSTATS
